Add active link highlighting for MultiNav (typical product pages) and refactor tab-navigation.js → move setActiveLink function to franklin lib
